exercise_4/frontend_py/requirements.txt
streamlit>=1.37,<2
requests>=2.31,<3
python-dotenv>=1.0,<2

